[#58] refactor : 기존 API에 로그인 반영 (3/3) (#66)

+++ src/main/java/mashup/backend/tich/category/controller/AdminCategoryController.java
@@ -7,15 +7,12 @@ import mashup.backend.tich.category.dto.CategorySaveRequestDto;
 import mashup.backend.tich.category.dto.CategoryUpdateRequestDto;
 import mashup.backend.tich.category.service.AdminCategoryService;
 import mashup.backend.tich.category.service.CategoryService;
-import mashup.backend.tich.user.domain.Role;
-import mashup.backend.tich.user.domain.User;
-import mashup.backend.tich.user.domain.UserRepository;
-import mashup.backend.tich.user.service.UserService;
+import mashup.backend.tich.exception.NoAccessException;
+import mashup.backend.tich.user.service.AdminUserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.persistence.NoResultException;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -24,13 +21,14 @@ import java.util.List;
 public class AdminCategoryController {
 
     private final AdminCategoryService adminCategoryService;
+    private final AdminUserService adminUserService;
     private final CategoryService categoryService;
-    private final UserRepository userRepository; /* 삭제 예정(관리자 확인용) */
-    private final UserService userService;
 
     @ApiOperation("카테고리 목록 조회")
     @GetMapping
-    public ResponseEntity<List<CategoryResponseDto>> showCategories() {
+    public ResponseEntity<List<CategoryResponseDto>> showCategories(@RequestHeader("TICH-TOKEN") String token) {
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
+
         List<CategoryResponseDto> categoryResponseDto = categoryService.showCategories();
 
         return ResponseEntity.status(HttpStatus.OK).body(categoryResponseDto);
@@ -38,7 +36,10 @@ public class AdminCategoryController {
 
     @ApiOperation("카테고리 상세 조회")
     @GetMapping("/{categoryId}")
-    public ResponseEntity<CategoryResponseDto> showCategory(@PathVariable Long categoryId) {
+    public ResponseEntity<CategoryResponseDto> showCategory(@RequestHeader("TICH-TOKEN") String token,
+                                                            @PathVariable Long categoryId) {
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
+
         CategoryResponseDto categoryResponseDto = categoryService.showCategory(categoryId);
 
         return ResponseEntity.status(HttpStatus.OK).body(categoryResponseDto);
@@ -46,12 +47,9 @@ public class AdminCategoryController {
 
     @ApiOperation("카테고리 등록")
     @PostMapping
-    public ResponseEntity<CategoryResponseDto> saveCategory(@RequestHeader("TICH-TOKEN") String accessToken,
+    public ResponseEntity<CategoryResponseDto> saveCategory(@RequestHeader("TICH-TOKEN") String token,
                                                             @RequestBody CategorySaveRequestDto requestDto) {
-//        // 임시 코드 : 추후 수정
-//        User user = makeTempUser();
-//        // ToDo : user check (accessToken)
-        User user = userService.findUserByToken(accessToken);
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         CategoryResponseDto categoryResponseDto = adminCategoryService.saveCategory(requestDto);
 
@@ -60,11 +58,9 @@ public class AdminCategoryController {
 
     @ApiOperation("카테고리 수정")
     @PutMapping
-    public ResponseEntity<CategoryResponseDto> updateCategory(@RequestHeader String accessToken,
+    public ResponseEntity<CategoryResponseDto> updateCategory(@RequestHeader("TICH-TOKEN") String token,
                                                               @RequestBody CategoryUpdateRequestDto requestDto) {
-        // 임시 코드 : 추후 수정
-        User user = makeTempUser();
-        // ToDo : user check (accessToken)
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         CategoryResponseDto categoryResponseDto = adminCategoryService.updateCategory(requestDto);
 
@@ -73,31 +69,12 @@ public class AdminCategoryController {
 
     @ApiOperation("카테고리 삭제")
     @DeleteMapping("/{categoryId}")
-    public ResponseEntity<Void> deleteCategory(@RequestHeader String accessToken,
+    public ResponseEntity<Void> deleteCategory(@RequestHeader("TICH-TOKEN") String token,
                                                @PathVariable Long categoryId) {
-        // 임시 코드 : 추후 수정
-        User user = makeTempUser();
-        // ToDo : user check (accessToken)
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         adminCategoryService.deleteCategory(categoryId);
 
         return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    /* 임시 코드 : 삭제 예정 */
-    private User makeTempUser() {
-        User user;
-        try {
-            user = userRepository.findByEmail("admin")
-                    .orElseThrow(() -> new NoResultException());
-        } catch (NoResultException e) {
-            user = userRepository.save(User.builder()
-                    .name("관리자")
-                    .email("admin")
-                    .picture("temp")
-                    .role(Role.ADMIN)
-                    .build());
-        }
-        return user;
     }
 }
